+++ CHANGELOG.md
@@ -55,6 +55,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - HTML rendering of Frames inside a Jupyter notebook.
 - set-theoretic functions: `union`, `intersect`, `setdiff` and `symdiff`.
 - support for multi-column keys.
+- ability to join Frames on multiple columns.
 
 #### Changed
 - `names` argument in `Frame()` constructor can no longer be a string --